Splitted view model. So now there 2 additional classes that are used -- JsonManager (for parsing, adding to and deleting from json) and Session (to start new translation, receiving bitmap and cancel)
